requirements.txt
bleach==2.1.2
certifi==2017.4.17
chardet==3.0.4
colorama==0.3.9
decorator==4.2.1
Django==2.0.2
django-autocomplete-light==3.3.0rc3
django-crispy-forms==1.7.2
django-extensions==1.9.9
django-filter==1.0.4
django-rest-framework==0.1.0
django-tables2==1.9.0
djangorestframework==3.7.7
entrypoints==0.2.3
html5lib==1.0.1
idna==2.5
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
lxml==3.8.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
pandocfilters==1.4.2
parso==0.1.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
pywinpty==0.5.1
pyzmq==16.0.4
rdflib==4.2.2
requests==2.18.1
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.10.0
SPARQLWrapper==1.8.0
style==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
typing==3.6.4
update==0.0.1
urllib3==1.21.1
wcwidth==0.1.7
webencodings==0.5.1
win-unicode-console==0.5
wincertstore==0.2
